test: introduce new test info pkg
