+++ test/hclutils/utils.go
@@ -20,8 +20,6 @@ import (
 
 	hhcl "github.com/hashicorp/hcl/v2"
 	"github.com/mineiros-io/terramate/errors"
-	"github.com/mineiros-io/terramate/hcl"
-	"github.com/mineiros-io/terramate/hcl/info"
 )
 
 // FixupFiledirOnErrorsFileRanges fix the filename in the ranges of the error list.
@@ -30,33 +28,6 @@ func FixupFiledirOnErrorsFileRanges(dir string, errs []error) {
 		if e, ok := err.(*errors.Error); ok {
 			e.FileRange.Filename = filepath.Join(dir, e.FileRange.Filename)
 		}
-	}
-}
-
-// FixRangesOnConfig fix the ranges on the given HCL config.
-// This is necessary since on tests we don't know the sandbox project
-// path, so host absolute paths must be updated here.
-func FixRangesOnConfig(dir string, cfg hcl.Config) {
-	for i := range cfg.Asserts {
-		cfg.Asserts[i].Range = newRange(dir, cfg.Asserts[i].Range)
-	}
-	for i := range cfg.Generate.Files {
-		cfg.Generate.Files[i].Range = newRange(dir,
-			cfg.Generate.Files[i].Range)
-
-		fixRangeOnAsserts(dir, cfg.Generate.Files[i].Asserts)
-	}
-	for i := range cfg.Generate.HCLs {
-		cfg.Generate.HCLs[i].Range = newRange(dir,
-			cfg.Generate.HCLs[i].Range)
-
-		fixRangeOnAsserts(dir, cfg.Generate.HCLs[i].Asserts)
-	}
-}
-
-func fixRangeOnAsserts(dir string, asserts []hcl.AssertConfig) {
-	for i := range asserts {
-		asserts[i].Range = newRange(dir, asserts[i].Range)
 	}
 }
 
@@ -80,19 +51,3 @@ func Start(line, column, char int) hhcl.Pos {
 
 // End pos of a range.
 func End(line, column, char int) hhcl.Pos { return Start(line, column, char) }
-
-func newRange(rootdir string, old info.Range) info.Range {
-	// When defining test cases there is no way to know the final
-	// absolute paths since sandboxes are dynamic/temporary.
-	// So we use relative paths as host paths and make them absolute here.
-	var zero info.Range
-	if old == zero {
-		// ast.Range is a zero value ast.Range, nothing to do
-		// avoid panics since the paths are not valid (empty strings).
-		return old
-	}
-	filename := filepath.Join(rootdir, old.HostPath())
-	return info.NewRange(rootdir, Mkrange(filename,
-		Start(old.Start().Line(), old.Start().Column(), old.Start().Byte()),
-		End(old.End().Line(), old.End().Column(), old.End().Byte())))
-}
